rename gloogateway to kgateway
