+++ test/kubernetes/e2e/tests/disabled_transformation_validation_test.go
@@ -11,7 +11,7 @@ import (
 	"github.com/kgateway-dev/kgateway/pkg/utils/envutils"
 	"github.com/kgateway-dev/kgateway/test/kubernetes/e2e"
 	. "github.com/kgateway-dev/kgateway/test/kubernetes/e2e/tests"
-	"github.com/kgateway-dev/kgateway/test/kubernetes/testutils/gloogateway"
+	"github.com/kgateway-dev/kgateway/test/kubernetes/testutils/kgateway"
 	"github.com/kgateway-dev/kgateway/test/testutils"
 )
 
@@ -22,7 +22,7 @@ func TestTransformationValidationDisabled(t *testing.T) {
 	installNs, nsEnvPredefined := envutils.LookupOrDefault(testutils.InstallNamespace, "disable-transformation-validation-test")
 	testInstallation := e2e.CreateTestInstallation(
 		t,
-		&gloogateway.Context{
+		&kgateway.Context{
 			InstallNamespace:          installNs,
 			ProfileValuesManifestFile: e2e.EdgeGatewayProfilePath,
 			ValuesManifestFile:        e2e.ManifestPath("disable-transformation-validation-helm.yaml"),
