+++ test/kubernetes/testutils/assertions/provider.go
@@ -10,7 +10,7 @@ import (
 	"github.com/stretchr/testify/require"
 
 	"github.com/kgateway-dev/kgateway/test/kubernetes/testutils/cluster"
-	"github.com/kgateway-dev/kgateway/test/kubernetes/testutils/gloogateway"
+	"github.com/kgateway-dev/kgateway/test/kubernetes/testutils/kgateway"
 )
 
 // Provider is the entity that provides methods which assert behaviors of a Kubernetes Cluster
@@ -24,7 +24,7 @@ type Provider struct {
 	Gomega gomega.Gomega
 
 	clusterContext     *cluster.Context
-	glooGatewayContext *gloogateway.Context
+	glooGatewayContext *kgateway.Context
 }
 
 // NewProvider returns a Provider that will provide Assertions that can be executed against an
@@ -48,7 +48,7 @@ func (p *Provider) WithClusterContext(clusterContext *cluster.Context) *Provider
 }
 
 // WithGlooGatewayContext sets the providers to point to a particular installation of Gloo Gateway
-func (p *Provider) WithGlooGatewayContext(ggCtx *gloogateway.Context) *Provider {
+func (p *Provider) WithGlooGatewayContext(ggCtx *kgateway.Context) *Provider {
 	p.glooGatewayContext = ggCtx
 	return p
 }
